requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
addict=2.4.0=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py36h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.26=h06a4308_2
cairo=1.16.0=hf32fb01_1
certifi=2021.5.30=py36h06a4308_0
cffi=1.14.6=py36h400218f_0
chardet=3.0.4=py36h06a4308_1003
cloudpickle=2.0.0=pyhd3eb1b0_0
cryptography=35.0.0=py36hd23ed53_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py36_0
cytoolz=0.11.0=py36h7b6447c_0
dask-core=2021.3.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=pyhd3eb1b0_0
expat=2.4.4=h295c915_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.1=h6c09931_0
freeglut=3.0.0=hf484d3e_5
freetype=2.11.0=h70c0345_0
future=0.18.2=py36_1
geos=3.8.0=he6710b0_0
glib=2.69.1=h4ff587b_1
graphite2=1.3.14=h23475e2_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=he6710b0_3
idna=2.10=pyhd3eb1b0_0
imageio=2.9.0=pyhd3eb1b0_0
imgaug=0.4.0=pyhd3eb1b0_0
intel-openmp=2022.0.1=h06a4308_3633
jasper=2.0.14=hd8c5072_2
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
jsonpatch=1.28=pyhd3eb1b0_0
jsonpointer=2.1=pyhd3eb1b0_0
kiwisolver=1.3.1=py36h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libglu=9.0.0=hf484d3e_1
libgomp=9.3.0=h5101ec6_17
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h7f8727e_2
libvpx=1.7.0=h439df22_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
matplotlib=3.3.4=py36h06a4308_0
matplotlib-base=3.3.4=py36h62a2d02_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
mmcv=1.4.8=pypi_0
ncurses=6.3=h7f8727e_2
networkx=2.5.1=pyhd3eb1b0_0
ninja=1.10.2=h5e70eb0_2
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
olefile=0.46=py36_0
opencv=3.4.2=py36h6fd60c2_1
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1m=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.1.5=py36ha9443f7_0
pcre=8.45=h295c915_0
pillow=8.3.1=py36h2c7a002_0
pip=21.2.2=py36h06a4308_0
pixman=0.40.0=h7f8727e_1
ptflops=0.6.8=pypi_0
py-opencv=3.4.2=py36hb342d67_1
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py36h05f1152_2
pysocks=1.7.1=py36h06a4308_0
python=3.6.12=hcff3b4d_2
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.6.0=py3.6_cuda10.2.89_cudnn7.6.5_0
pytz=2021.3=pyhd3eb1b0_0
pywavelets=1.1.1=py36h7b6447c_2
pyyaml=5.4.1=py36h27cfd23_1
pyzmq=20.0.0=py36h2531618_1
qt=5.9.7=h5867ecd_1
readline=8.1.2=h7f8727e_1
requests=2.25.0=pyhd3eb1b0_0
scikit-image=0.16.2=py36h0573a6f_0
scikit-learn=0.24.1=py36ha9443f7_0
scipy=1.5.2=py36h0b6359f_0
setuptools=58.0.4=py36h06a4308_0
shapely=1.7.1=py36h1728cc4_0
sip=4.19.8=py36hf484d3e_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.37.2=hc218d9a_0
tabulate=0.8.7=py36h06a4308_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h1ccaba5_0
toolz=0.11.2=pyhd3eb1b0_0
torchvision=0.7.0=py36_cu102
tornado=6.1=py36h27cfd23_0
urllib3=1.25.11=py_0
websocket-client=0.57.0=py36_2
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yapf=0.32.0=pypi_0
zeromq=4.3.4=h2531618_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
